'Lazy parsing' replaced by 'parsing by lines' + safer code
Change of strategy to parse a large XML: instead of parsing it lazyly do:
  (*) read the file and populate an RDD with its lines
  (*) 'map' onto the RDD to convert each line in a usable, parsed structure.

This allows to work solely on RDDs, delaying as much as possible the evaluation of the actual contents - which are the contents eventually saved into a file, so potentially optimized for minimal memory footprint.

The code also has been rendered safer by allowing the compiler to help with problems (by specifying more pointed types and sharing structures on access).